Moved gp-v1 api file from /api folder to app's root
