+++ src/components/DepositWidget/useRowActions.tsx
@@ -10,7 +10,7 @@ import { formatSmart, formatAmountFull, logDebug, getToken, safeFilledToken } fr
 import { composeOptionalParams } from 'utils/transaction'
 
 // Api
-import { depositApi, erc20Api } from 'api/gpV1'
+import { depositApi, erc20Api } from 'apps/gp-v1/api'
 
 // Hooks and reducers
 import { useWalletConnection } from 'hooks/useWalletConnection'
