+++ src/components/TradeWidget/index.tsx
@@ -28,7 +28,7 @@ import {
 
 // api
 import { PendingTxObj } from 'api/exchange/ExchangeApi'
-import { tokenListApi } from 'api/gpV1'
+import { tokenListApi } from 'apps/gp-v1/api'
 
 // components
 import OrdersWidget from 'components/OrdersWidget'
